+++ app/javascript/vue/components/georeferences/icons/index.js
@@ -1,71 +1,38 @@
 import L from 'leaflet'
-
-import blueIcon from './img/marker-icon-2x-blue.png'
-import goldIcon from './img/marker-icon-2x-gold.png'
-import redIcon from './img/marker-icon-2x-red.png'
-import greenIcon from './img/marker-icon-2x-green.png'
-import orangeIcon from './img/marker-icon-2x-orange.png'
-import yellowIcon from './img/marker-icon-2x-yellow.png'
-import violetIcon from './img/marker-icon-2x-violet.png'
-import greyIcon from './img/marker-icon-2x-grey.png'
-import blackIcon from './img/marker-icon-2x-black.png'
-import shadowImage from './img/marker-shadow.png'
+import {
+  COLLECTION_OBJECT,
+  FIELD_OCCURRENCE,
+  TYPE_MATERIAL,
+  ASSERTED_DISTRIBUTION,
+  GEOREFERENCE
+} from '@/constants'
 
 const defaultProperties = {
-  shadowUrl: shadowImage,
-  iconSize: [25, 41],
-  iconAnchor: [12, 41],
-  popupAnchor: [1, -34],
-  shadowSize: [41, 41]
+  iconSize: [8, 8],
+  iconAnchor: [4, 4]
 }
 
 const Icon = {
-  blue: new L.Icon({
-    iconUrl: blueIcon,
+  [COLLECTION_OBJECT]: {
+    className: 'map-point-marker bg-collection-object',
     ...defaultProperties
-  }),
-
-  gold: new L.Icon({
-    iconUrl: goldIcon,
+  },
+  [FIELD_OCCURRENCE]: {
+    className: 'map-point-marker bg-field-occurrence',
     ...defaultProperties
-  }),
-
-  red: new L.Icon({
-    iconUrl: redIcon,
+  },
+  [TYPE_MATERIAL]: {
+    className: 'map-point-marker bg-type-material',
     ...defaultProperties
-  }),
-
-  green: new L.Icon({
-    iconUrl: greenIcon,
+  },
+  [ASSERTED_DISTRIBUTION]: {
+    className: 'map-point-marker bg-asserted-distribution',
     ...defaultProperties
-  }),
-
-  orange: new L.Icon({
-    iconUrl: orangeIcon,
+  },
+  [GEOREFERENCE]: {
+    className: 'map-point-marker bg-georeference',
     ...defaultProperties
-  }),
-
-  yellow: new L.Icon({
-    iconUrl: yellowIcon,
-    ...defaultProperties
-  }),
-
-  violet: new L.Icon({
-    iconUrl: violetIcon,
-    ...defaultProperties
-  }),
-
-  grey: new L.Icon({
-    iconUrl: greyIcon,
-    ...defaultProperties
-  }),
-
-  black: new L.Icon({
-    iconUrl: blackIcon,
-    ...defaultProperties
-  })
+  }
 }
 
-export {
-  Icon
-}
+export { Icon }