+++ app/models/project.rb
@@ -52,6 +52,7 @@ class Project < ApplicationRecord
      Protocol
      SqedDepiction
      Depiction
+     Conveyance
      SledImage
      Documentation
      Document
@@ -81,6 +82,7 @@ class Project < ApplicationRecord
      TaxonDetermination
      TypeMaterial
      Image
+     Sound
      CommonName
      TaxonNameClassification
      TaxonNameRelationship